+++ src/wetterdienst/provider/dwd/observation/api.py
@@ -6,18 +6,29 @@ from __future__ import annotations
 
 import datetime as dt
 import logging
-from collections.abc import Iterable
+from collections import defaultdict
+from collections.abc import Iterable, Iterator
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, ClassVar, Literal
 from zoneinfo import ZoneInfo
 
 import polars as pl
 import portion
+from fsspec.implementations.zip import ZipFileSystem
 from portion import Interval
 
 from wetterdienst.metadata.cache import CacheExpiry
 from wetterdienst.metadata.period import Period
 from wetterdienst.metadata.resolution import Resolution
+from wetterdienst.model.history import (
+    History,
+    TimeseriesHistory,
+    _DeviceHistory,
+    _GeographyHistory,
+    _MissingDataHistory,
+    _NameHistory,
+    _ParameterHistory,
+)
 from wetterdienst.model.metadata import DatasetModel, ParameterSearch
 from wetterdienst.model.request import TimeseriesRequest
 from wetterdienst.model.values import TimeseriesValues
@@ -26,6 +37,7 @@ from wetterdienst.provider.dwd.observation.download import (
 )
 from wetterdienst.provider.dwd.observation.fileindex import (
     _build_url_from_dataset_and_period,
+    _create_file_index_for_1minute_or_5minute_historical_precipitation_metadata,
     _create_file_index_for_dwd_server,
     create_file_index_for_climate_observations,
     create_file_list_for_climate_observations,
@@ -40,6 +52,7 @@ from wetterdienst.provider.dwd.observation.metaindex import (
 from wetterdienst.provider.dwd.observation.parser import parse_climate_observations_data
 from wetterdienst.settings import Settings
 from wetterdienst.util.enumeration import parse_enumeration_from_template
+from wetterdienst.util.network import File, download_file
 from wetterdienst.util.python import to_list
 
 if TYPE_CHECKING:
@@ -188,12 +201,308 @@ class DwdObservationValues(TimeseriesValues):
         return file_index.collect().get_column("date_range").to_list()
 
 
+class DwdObservationHistory(TimeseriesHistory):
+    """History class for DWD observation data."""
+
+    def _collect_station_history(self, station_id: str, available_datasets: list[DatasetModel]) -> Iterator[History]:
+        for dataset in available_datasets:
+            dataset_identifier = f"{dataset.resolution.value.name}/{dataset.name}/{station_id}"
+            log.info(f"Acquiring station history for {dataset_identifier}.")
+            try:
+                if dataset.resolution.value in HIGH_RESOLUTIONS:
+                    file_index = _create_file_index_for_1minute_or_5minute_historical_precipitation_metadata(
+                        resolution=dataset.resolution.value.value,
+                        dataset=dataset.name_original,
+                        settings=self.sr.stations.settings,
+                    )
+                else:
+                    file_index = create_file_index_for_climate_observations(
+                        dataset=dataset,
+                        period=Period.HISTORICAL,
+                        settings=self.sr.stations.settings,
+                    )
+            except FileNotFoundError:
+                log.info(
+                    f"No files found for {dataset_identifier} and period {Period.HISTORICAL}. Station will be skipped."
+                )
+                continue
+            file_index = file_index.filter(pl.col("station_id").eq(station_id)).collect()
+            if file_index.is_empty():
+                log.info(f"No files found for {dataset_identifier}. Station will be skipped.")
+                continue
+            if dataset.resolution.value in (Resolution.SUBDAILY, Resolution.DAILY) and dataset.name == "wind_extreme":
+                # subdaily extreme wind dataset has one file for FX3 and one for FX6
+                name_histories = []
+                parameter_histories = []
+                device_histories = []
+                geography_histories = []
+                missing_data_histories = []
+
+                for file in file_index.get_column("url"):
+                    file: File = download_file(
+                        url=file,
+                        cache_dir=self.sr.stations.settings.cache_dir,
+                        ttl=CacheExpiry.METAINDEX,
+                        client_kwargs=self.sr.stations.settings.fsspec_client_kwargs,
+                        cache_disable=self.sr.stations.settings.cache_disable,
+                    )
+                    name_history = self.read_name_history(file)
+                    name_histories.append(name_history)
+                    parameter_history_list = self.read_parameter_history(file)
+                    parameter_histories.append(parameter_history_list)
+                    device_history_list = self.read_device_history(file)
+                    device_histories.append(device_history_list)
+                    geography_history_list = self.read_geography_history(file)
+                    geography_histories.append(geography_history_list)
+                    missing_data_history = self.read_missing_data_history(file, dataset.resolution.value)
+                    missing_data_histories.append(missing_data_history)
+                yield History(
+                    name=_NameHistory(
+                        station=[station for nh in name_histories for station in nh.station],
+                        operator=[operator for nh in name_histories for operator in nh.operator],
+                    ),
+                    parameter=[ph for sublist in parameter_histories for ph in sublist],
+                    device=[dh for sublist in device_histories for dh in sublist],
+                    geography=[gh for sublist in geography_histories for gh in sublist],
+                    missing_data=_MissingDataHistory(
+                        summary=[summary for mdh in missing_data_histories for summary in mdh.summary],
+                        periods=[periods for mdh in missing_data_histories for periods in mdh.periods],
+                    ),
+                )
+            file: File = download_file(
+                url=file_index.get_column("url").item(),
+                cache_dir=self.sr.stations.settings.cache_dir,
+                ttl=CacheExpiry.METAINDEX,
+                client_kwargs=self.sr.stations.settings.fsspec_client_kwargs,
+                cache_disable=self.sr.stations.settings.cache_disable,
+            )
+            name_history = self.read_name_history(file)
+            parameter_history_list = self.read_parameter_history(file)
+            device_history_list = self.read_device_history(file)
+            geography_history_list = self.read_geography_history(file)
+            missing_data_history = self.read_missing_data_history(file, dataset.resolution.value)
+            yield History(
+                name=name_history,
+                parameter=parameter_history_list,
+                device=device_history_list,
+                geography=geography_history_list,
+                missing_data=missing_data_history,
+            )
+
+    @staticmethod
+    def read_name_history(file: File) -> _NameHistory:
+        """Read station name and operator name history from a zip file."""
+        zfs = ZipFileSystem(file.content)
+        # find file
+        name_file = zfs.glob("Metadaten_Stationsname_Betreibername_*.txt")
+        if not name_file:
+            return _NameHistory()
+        name_file = name_file[0]
+        with zfs.open(name_file, "r") as f:
+            lines = f.readlines()
+        # split by empty line into station name and operator name
+        split_index = lines.index("\n")
+        # determine which part is station name and which is operator name
+        if "Stationsname" in lines[0]:
+            station_name_lines = lines[1:split_index]
+            operator_name_lines = lines[split_index + 2 : -1]
+            if operator_name_lines[-1].startswith("generiert:"):
+                operator_name_lines = operator_name_lines[:-1]
+        else:
+            operator_name_lines = lines[1:split_index]
+            station_name_lines = lines[split_index + 2 : -1]
+            if station_name_lines[-1].startswith("generiert:"):
+                station_name_lines = station_name_lines[:-1]
+
+        station_name_data = []
+        for line in station_name_lines:
+            parts = line.strip().split(";")
+            station_name_data.append(
+                {
+                    "station_id": parts[0].strip().zfill(5),
+                    "station_name": parts[1].strip(),
+                    "start_date": dt.datetime.strptime(parts[2].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC")),
+                    "end_date": (
+                        dt.datetime.strptime(parts[3].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC"))
+                        if parts[3].strip()
+                        else None
+                    ),
+                }
+            )
+        operator_name_data = []
+        for line in operator_name_lines:
+            parts = line.strip().split(";")
+            operator_name_data.append(
+                {
+                    "station_id": parts[0].strip().zfill(5),
+                    "operator_name": parts[1].strip(),
+                    "start_date": dt.datetime.strptime(parts[2].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC")),
+                    "end_date": (
+                        dt.datetime.strptime(parts[3].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC"))
+                        if parts[3].strip()
+                        else None
+                    ),
+                }
+            )
+        return _NameHistory(station=station_name_data, operator=operator_name_data)
+
+    @staticmethod
+    def read_parameter_history(file: File) -> list[_ParameterHistory]:
+        """Read parameter history from file."""
+        zfs = ZipFileSystem(file.content)
+        # find file Metadaten_Parameter_klima_tag_01048.txt, but use glob to be independent of station
+        parameter_file = zfs.glob("**/Metadaten_Parameter_*.txt")
+        if not parameter_file:
+            return []
+        parameter_file = parameter_file[0]
+        parameter_text = zfs.open(parameter_file).read()
+        lines = parameter_text.strip().decode(encoding="latin1").splitlines()
+        records = []
+        for line in lines[1:]:
+            if line.startswith(("Legende:", "generiert:")):
+                continue
+            parts = line.split(";")
+            if len(parts) < 8:
+                continue
+            record = {
+                "station_id": parts[0].strip(),
+                "start_date": dt.datetime.strptime(parts[1].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC")),
+                "end_date": dt.datetime.strptime(parts[2].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC")),
+                "station_name": parts[3].strip(),
+                "parameter": parts[4].strip(),
+                "description": parts[5].strip(),
+                "unit": parts[6].strip(),
+                "data_source": parts[7].strip(),
+                "extra_info": parts[8].strip(),
+                "special": parts[9].strip(),
+                "literature": parts[10].strip(),
+            }
+            records.append(record)
+        return [_ParameterHistory.model_validate(record) for record in records]
+
+    @staticmethod
+    def read_device_history(file: File) -> list[_DeviceHistory]:
+        """Read device history from file."""
+        zfs = ZipFileSystem(file.content)
+        # find all Metadaten_Geraete_*.txt files, but use glob to be independent of station
+        device_files = zfs.glob("**/Metadaten_Geraete_*.txt")
+        records = []
+        for device_file in device_files:
+            device_text = zfs.open(device_file).read()
+            lines = device_text.strip().decode(encoding="latin1").splitlines()
+            for line in lines[1:]:
+                if line.startswith("generiert:"):
+                    continue
+                parts = line.split(";")
+                if len(parts) < 10:
+                    continue
+                record = {
+                    "station_id": parts[0].strip(),
+                    "station_name": parts[1].strip(),
+                    "longitude": parts[2].strip() if parts[2].strip() else None,
+                    "latitude": parts[3].strip() if parts[3].strip() else None,
+                    "station_height": parts[4].strip() if parts[4].strip() else None,
+                    "device_height": parts[5].strip() if parts[5].strip() else None,
+                    "start_date": dt.datetime.strptime(parts[6].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC")),
+                    "end_date": dt.datetime.strptime(parts[7].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC")),
+                    "device_type": parts[8].strip(),
+                    "method": parts[9].strip(),
+                }
+                records.append(record)
+        return [_DeviceHistory.model_validate(record) for record in records]
+
+    @staticmethod
+    def read_geography_history(file: File) -> list[_GeographyHistory]:
+        """Read geography history from file."""
+        zfs = ZipFileSystem(file.content)
+        # find all Metadaten_Geographie_*.txt files, but use glob to be independent of station
+        geography_files = zfs.glob("**/Metadaten_Geographie_*.txt")
+        records = []
+        for geography_file in geography_files:
+            geography_text = zfs.open(geography_file).read()
+            lines = geography_text.strip().decode(encoding="latin1").splitlines()
+            for line in lines[1:]:
+                parts = line.split(";")
+                if len(parts) < 7:
+                    continue
+                record = {
+                    "station_id": parts[0].strip(),
+                    "station_height": float(parts[1].strip()),
+                    "latitude": float(parts[2].strip()),
+                    "longitude": float(parts[3].strip()),
+                    "start_date": dt.datetime.strptime(parts[4].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC")),
+                    "end_date": dt.datetime.strptime(parts[5].strip(), "%Y%m%d").replace(tzinfo=ZoneInfo("UTC"))
+                    if parts[5].strip()
+                    else dt.datetime.now(tz=ZoneInfo("UTC")),
+                    "station_name": parts[6].strip(),
+                }
+                records.append(record)
+        return [_GeographyHistory.model_validate(record) for record in records]
+
+    @staticmethod
+    def read_missing_data_history(file: File, resolution: Resolution) -> _MissingDataHistory:
+        """Read missing data history from file."""
+        # we need to adjust date format based on resolution for period records
+        resolution_to_date_format = defaultdict(
+            lambda: "%d.%m.%Y", {Resolution.HOURLY: "%d.%m.%Y-%H:%M", Resolution.SUBDAILY: "%d.%m.%Y-%H:%M"}
+        )
+        date_format = resolution_to_date_format[resolution]
+        zfs = ZipFileSystem(file.content)
+        # find all Metadaten_Fehlzeiten_*.txt files, but use glob to be independent of station
+        missing_data_file = zfs.glob("**/Metadaten_Fehldaten_*.txt")
+        if not missing_data_file:
+            return _MissingDataHistory()
+        missing_data_file = missing_data_file[0]
+        missing_data_text = zfs.open(missing_data_file).read()
+        lines = missing_data_text.decode("utf-8").strip().splitlines()
+        # find line that separates summary from period
+        split_line = 0
+        count_starts_with = 0
+        for i, line in enumerate(lines):
+            if line.strip().startswith("Stations_ID"):
+                count_starts_with += 1
+                if count_starts_with == 2:
+                    split_line = i
+                    break
+        summary_lines = lines[1:split_line]
+        period_lines = lines[split_line + 1 : -1]
+        summary_records = []
+        for line in summary_lines:
+            parts = line.split(";")
+            record = {
+                "station_id": parts[0].zfill(5),
+                "station_name": parts[1],
+                "parameter": parts[2],
+                "start_date": dt.datetime.strptime(parts[3], "%d.%m.%Y").replace(tzinfo=ZoneInfo("UTC")),
+                "end_date": dt.datetime.strptime(parts[4], "%d.%m.%Y").replace(tzinfo=ZoneInfo("UTC")),
+                "missing_count": int(parts[5]) if parts[5] else None,
+                "description": parts[6],
+            }
+            summary_records.append(record)
+        # period records use date format depending on resolution
+        period_records = []
+        for line in period_lines:
+            parts = line.split(";")
+            record = {
+                "station_id": parts[0].zfill(5),
+                "station_name": parts[1],
+                "parameter": parts[2],
+                "start_date": dt.datetime.strptime(parts[3], date_format).replace(tzinfo=ZoneInfo("UTC")),
+                "end_date": dt.datetime.strptime(parts[4], date_format).replace(tzinfo=ZoneInfo("UTC")),
+                "missing_count": int(parts[5]) if parts[5] else None,
+                "description": parts[6],
+            }
+            period_records.append(record)
+        return _MissingDataHistory(summary=summary_records, periods=period_records)
+
+
 @dataclass
 class DwdObservationRequest(TimeseriesRequest):
     """Request class for DWD observation data."""
 
     metadata = DwdObservationMetadata
     _values = DwdObservationValues
+    _history = DwdObservationHistory
     _available_periods: ClassVar = {Period.HISTORICAL, Period.RECENT, Period.NOW}
     periods: str | Period | set[str | Period] = None